Deduplicate `EnvironmentVars`
