+++ cyclonedx-bom/src/specs/v1_5/formulation/workflow/input.rs
@@ -1,17 +1,16 @@
 use serde::{Deserialize, Serialize};
-use xml::{reader, writer};
+use xml::reader;
 
 use crate::{
     errors::XmlReadError,
     specs::{common::property::Properties, v1_5::attachment::Attachment},
     xml::{
-        attribute_or_error, read_lax_validation_tag, read_simple_tag, to_xml_read_error,
-        to_xml_write_error, unexpected_element_error, write_close_tag, write_simple_tag,
-        write_start_tag, FromXml, ToInnerXml, ToXml,
+        read_lax_validation_tag, read_simple_tag, to_xml_read_error, unexpected_element_error,
+        write_close_tag, write_simple_tag, write_start_tag, FromXml, ToInnerXml, ToXml,
     },
 };
 
-use super::resource_reference::ResourceReference;
+use super::{resource_reference::ResourceReference, EnvironmentVars, ENVIRONMENT_VARS_TAG};
 
 #[derive(Debug, PartialEq, Serialize, Deserialize)]
 pub(crate) struct Input {
@@ -315,105 +314,6 @@ impl FromXml for Parameter {
             value,
             data_type,
         })
-    }
-}
-
-#[derive(Debug, PartialEq, Serialize, Deserialize)]
-pub(crate) struct EnvironmentVars(pub(crate) Vec<EnvironmentVar>);
-
-const ENVIRONMENT_VARS_TAG: &str = "environmentVars";
-
-impl ToXml for EnvironmentVars {
-    fn write_xml_element<W: std::io::prelude::Write>(
-        &self,
-        writer: &mut xml::EventWriter<W>,
-    ) -> Result<(), crate::errors::XmlWriteError> {
-        write_start_tag(writer, ENVIRONMENT_VARS_TAG)?;
-
-        for environment_var in &self.0 {
-            environment_var.write_xml_element(writer)?;
-        }
-
-        write_close_tag(writer, ENVIRONMENT_VARS_TAG)
-    }
-}
-
-impl FromXml for EnvironmentVars {
-    fn read_xml_element<R: std::io::prelude::Read>(
-        event_reader: &mut xml::EventReader<R>,
-        element_name: &xml::name::OwnedName,
-        _attributes: &[xml::attribute::OwnedAttribute],
-    ) -> Result<Self, XmlReadError>
-    where
-        Self: Sized,
-    {
-        let mut environment_vars = vec![];
-
-        let mut got_end_tag = false;
-        while !got_end_tag {
-            let next_element = event_reader
-                .next()
-                .map_err(to_xml_read_error(ENVIRONMENT_VARS_TAG))?;
-            match next_element {
-                reader::XmlEvent::StartElement {
-                    name, attributes, ..
-                } if name.local_name == ENVIRONMENT_VAR_TAG => {
-                    let value = read_simple_tag(event_reader, &name)?;
-                    let name = attribute_or_error(&name, &attributes, NAME_ATTR)?;
-                    environment_vars.push(EnvironmentVar::Property { name, value });
-                }
-                reader::XmlEvent::StartElement { name, .. } if name.local_name == VALUE_TAG => {
-                    let value = read_simple_tag(event_reader, &name)?;
-                    environment_vars.push(EnvironmentVar::Value(value));
-                }
-                // lax validation of any elements from a different schema
-                reader::XmlEvent::StartElement { name, .. } => {
-                    read_lax_validation_tag(event_reader, &name)?
-                }
-                reader::XmlEvent::EndElement { name } if &name == element_name => {
-                    got_end_tag = true;
-                }
-                unexpected => return Err(unexpected_element_error(element_name, unexpected)),
-            }
-        }
-
-        Ok(Self(environment_vars))
-    }
-}
-
-#[derive(Debug, PartialEq, Serialize, Deserialize)]
-#[serde(untagged, rename_all = "camelCase")]
-pub(crate) enum EnvironmentVar {
-    Property { name: String, value: String },
-    Value(String),
-}
-
-const ENVIRONMENT_VAR_TAG: &str = "environmentVar";
-const NAME_ATTR: &str = "name";
-
-impl ToXml for EnvironmentVar {
-    fn write_xml_element<W: std::io::prelude::Write>(
-        &self,
-        writer: &mut xml::EventWriter<W>,
-    ) -> Result<(), crate::errors::XmlWriteError> {
-        match self {
-            Self::Property { name, value } => {
-                writer
-                    .write(writer::XmlEvent::start_element(ENVIRONMENT_VAR_TAG).attr("name", name))
-                    .map_err(to_xml_write_error(ENVIRONMENT_VAR_TAG))?;
-
-                writer
-                    .write(writer::XmlEvent::characters(value))
-                    .map_err(to_xml_write_error(ENVIRONMENT_VAR_TAG))?;
-
-                write_close_tag(writer, ENVIRONMENT_VAR_TAG)?;
-            }
-            Self::Value(value) => {
-                write_simple_tag(writer, VALUE_TAG, value)?;
-            }
-        }
-
-        Ok(())
     }
 }
