+++ cyclonedx-bom/src/specs/v1_5/formulation/workflow/mod.rs
@@ -1,11 +1,20 @@
+mod input;
+mod output;
+mod resource_reference;
+mod steps;
+mod trigger;
+mod workspace;
+
 use crate::{
     elem_tag,
     errors::XmlReadError,
     get_elements_lax,
     specs::common::{bom_reference::BomReference, dependency::Dependency, property::Properties},
     xml::{
-        attribute_or_error, read_simple_tag, to_xml_write_error, write_close_tag, write_list_tag,
-        write_simple_option_tag, write_simple_tag, FromXml, ToXml, VecElemTag, VecXmlReader,
+        attribute_or_error, read_lax_validation_tag, read_simple_tag, to_xml_read_error,
+        to_xml_write_error, unexpected_element_error, write_close_tag, write_list_tag,
+        write_simple_option_tag, write_simple_tag, write_start_tag, FromXml, ToXml, VecElemTag,
+        VecXmlReader,
     },
 };
 
@@ -14,15 +23,8 @@ use self::{
     trigger::Trigger, workspace::Workspace,
 };
 
-mod input;
-mod output;
-mod resource_reference;
-mod steps;
-mod trigger;
-mod workspace;
-
 use serde::{Deserialize, Serialize};
-use xml::writer;
+use xml::{reader, writer};
 
 use super::runtime_topology::RuntimeTopology;
 
@@ -329,6 +331,106 @@ impl FromXml for Task {
     }
 }
 
+#[derive(Debug, PartialEq, Serialize, Deserialize)]
+pub(crate) struct EnvironmentVars(pub(crate) Vec<EnvironmentVar>);
+
+const ENVIRONMENT_VARS_TAG: &str = "environmentVars";
+
+impl ToXml for EnvironmentVars {
+    fn write_xml_element<W: std::io::prelude::Write>(
+        &self,
+        writer: &mut xml::EventWriter<W>,
+    ) -> Result<(), crate::errors::XmlWriteError> {
+        write_start_tag(writer, ENVIRONMENT_VARS_TAG)?;
+
+        for environment_var in &self.0 {
+            environment_var.write_xml_element(writer)?;
+        }
+
+        write_close_tag(writer, ENVIRONMENT_VARS_TAG)
+    }
+}
+
+impl FromXml for EnvironmentVars {
+    fn read_xml_element<R: std::io::prelude::Read>(
+        event_reader: &mut xml::EventReader<R>,
+        element_name: &xml::name::OwnedName,
+        _attributes: &[xml::attribute::OwnedAttribute],
+    ) -> Result<Self, XmlReadError>
+    where
+        Self: Sized,
+    {
+        let mut environment_vars = vec![];
+
+        let mut got_end_tag = false;
+        while !got_end_tag {
+            let next_element = event_reader
+                .next()
+                .map_err(to_xml_read_error(ENVIRONMENT_VARS_TAG))?;
+            match next_element {
+                reader::XmlEvent::StartElement {
+                    name, attributes, ..
+                } if name.local_name == ENVIRONMENT_VAR_TAG => {
+                    let value = read_simple_tag(event_reader, &name)?;
+                    let name = attribute_or_error(&name, &attributes, NAME_ATTR)?;
+                    environment_vars.push(EnvironmentVar::Property { name, value });
+                }
+                reader::XmlEvent::StartElement { name, .. } if name.local_name == VALUE_TAG => {
+                    let value = read_simple_tag(event_reader, &name)?;
+                    environment_vars.push(EnvironmentVar::Value(value));
+                }
+                // lax validation of any elements from a different schema
+                reader::XmlEvent::StartElement { name, .. } => {
+                    read_lax_validation_tag(event_reader, &name)?
+                }
+                reader::XmlEvent::EndElement { name } if &name == element_name => {
+                    got_end_tag = true;
+                }
+                unexpected => return Err(unexpected_element_error(element_name, unexpected)),
+            }
+        }
+
+        Ok(Self(environment_vars))
+    }
+}
+
+#[derive(Debug, PartialEq, Serialize, Deserialize)]
+#[serde(untagged, rename_all = "camelCase")]
+pub(crate) enum EnvironmentVar {
+    Property { name: String, value: String },
+    Value(String),
+}
+
+const ENVIRONMENT_VAR_TAG: &str = "environmentVar";
+const VALUE_TAG: &str = "value";
+const NAME_ATTR: &str = "name";
+
+impl ToXml for EnvironmentVar {
+    fn write_xml_element<W: std::io::prelude::Write>(
+        &self,
+        writer: &mut xml::EventWriter<W>,
+    ) -> Result<(), crate::errors::XmlWriteError> {
+        match self {
+            Self::Property { name, value } => {
+                writer
+                    .write(writer::XmlEvent::start_element(ENVIRONMENT_VAR_TAG).attr("name", name))
+                    .map_err(to_xml_write_error(ENVIRONMENT_VAR_TAG))?;
+
+                writer
+                    .write(writer::XmlEvent::characters(value))
+                    .map_err(to_xml_write_error(ENVIRONMENT_VAR_TAG))?;
+
+                write_close_tag(writer, ENVIRONMENT_VAR_TAG)?;
+            }
+            Self::Value(value) => {
+                write_simple_tag(writer, VALUE_TAG, value)?;
+            }
+        }
+
+        Ok(())
+    }
+}
+
 #[cfg(test)]
 mod test {
     use crate::{
@@ -523,12 +625,10 @@ mod test {
             inputs: Some(vec![
                 Input {
                     required: RequiredInputField::EnvironmentVars {
-                        environment_vars: input::EnvironmentVars(vec![
-                            input::EnvironmentVar::Property {
-                                name: "Foo".into(),
-                                value: "Bar".into(),
-                            },
-                        ]),
+                        environment_vars: EnvironmentVars(vec![EnvironmentVar::Property {
+                            name: "Foo".into(),
+                            value: "Bar".into(),
+                        }]),
                     },
                     source: None,
                     target: None,
@@ -536,9 +636,9 @@ mod test {
                 },
                 Input {
                     required: RequiredInputField::EnvironmentVars {
-                        environment_vars: input::EnvironmentVars(vec![
-                            input::EnvironmentVar::Value("FooBar".into()),
-                        ]),
+                        environment_vars: EnvironmentVars(vec![EnvironmentVar::Value(
+                            "FooBar".into(),
+                        )]),
                     },
                     source: None,
                     target: None,
@@ -546,12 +646,12 @@ mod test {
                 },
                 Input {
                     required: RequiredInputField::EnvironmentVars {
-                        environment_vars: input::EnvironmentVars(vec![
-                            input::EnvironmentVar::Property {
+                        environment_vars: EnvironmentVars(vec![
+                            EnvironmentVar::Property {
                                 name: "Foo".into(),
                                 value: "Bar".into(),
                             },
-                            input::EnvironmentVar::Value("FooBar".into()),
+                            EnvironmentVar::Value("FooBar".into()),
                         ]),
                     },
                     source: None,
@@ -562,7 +662,10 @@ mod test {
             outputs: Some(vec![
                 Output {
                     required: RequiredOutputField::EnvironmentVars {
-                        environment_vars: output::EnvironmentVars(vec![]),
+                        environment_vars: EnvironmentVars(vec![EnvironmentVar::Property {
+                            name: "Foo".into(),
+                            value: "Bar".into(),
+                        }]),
                     },
                     r#type: None,
                     source: None,
@@ -571,7 +674,9 @@ mod test {
                 },
                 Output {
                     required: RequiredOutputField::EnvironmentVars {
-                        environment_vars: output::EnvironmentVars(vec![]),
+                        environment_vars: EnvironmentVars(vec![EnvironmentVar::Value(
+                            "FooBar".into(),
+                        )]),
                     },
                     r#type: None,
                     source: None,
@@ -580,7 +685,13 @@ mod test {
                 },
                 Output {
                     required: RequiredOutputField::EnvironmentVars {
-                        environment_vars: output::EnvironmentVars(vec![]),
+                        environment_vars: EnvironmentVars(vec![
+                            EnvironmentVar::Property {
+                                name: "Foo".into(),
+                                value: "Bar".into(),
+                            },
+                            EnvironmentVar::Value("FooBar".into()),
+                        ]),
                     },
                     r#type: None,
                     source: None,
@@ -860,6 +971,6 @@ mod test {
 "#;
         let actual: Workflow = read_element_from_string(input);
         let expected = example_workflow();
-        assert_eq!(actual, expected);
+        pretty_assertions::assert_eq!(actual, expected);
     }
 }
